edit

+++ lib/pages/post/my_page.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:firebase_storage/firebase_storage.dart'; // Firebase Storage 추가
+import 'package:firebase_storage/firebase_storage.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import '../../firebase/login_google.dart';
 import '../../firebase/user_data.dart';
@@ -28,17 +28,17 @@ class MyPage extends StatefulWidget {
 class _MyPageState extends State<MyPage> {
   late UserModel user;
   final ImagePicker _picker = ImagePicker();
+  final String defaultImageUrl = 'assets/images/default_image.png';
+  bool isImageChanged = false;
 
   @override
   void initState() {
     super.initState();
-    // 사용자 정보 초기화
     user = UserModel(
       uid: FirebaseAuth.instance.currentUser!.uid,
-      profileImageUrl: '', // 초기 프로필 이미지 경로를 빈 문자열로 설정
-      nickname: '', // 초기 닉네임을 빈 문자열로 설정
+      profileImageUrl: defaultImageUrl,
+      nickname: '',
     );
-    // 사용자 정보 불러오기
     _loadUserData();
   }
 
@@ -51,7 +51,6 @@ class _MyPageState extends State<MyPage> {
           IconButton(
             icon: const Icon(Icons.logout),
             onPressed: () {
-              // 로그아웃 기능 실행
               _logout(context);
             },
           ),
@@ -66,9 +65,9 @@ class _MyPageState extends State<MyPage> {
                 _pickImage();
               },
               child: CircleAvatar(
-                backgroundImage: user.profileImageUrl.isEmpty
-                    ? Image.asset('assets/images/default_image.png').image
-                    : NetworkImage(user.profileImageUrl),
+                backgroundImage: user.profileImageUrl == defaultImageUrl
+                    ? Image.asset(defaultImageUrl).image
+                    : NetworkImage(user.profileImageUrl) as ImageProvider,
                 radius: 50,
               ),
             ),
@@ -80,10 +79,10 @@ class _MyPageState extends State<MyPage> {
                   return const CircularProgressIndicator();
                 }
                 if (!snapshot.hasData) {
-                  return Text(user.nickname, style: const TextStyle(fontSize: 20)); // 에러 발생 시 기본 닉네임 표시
+                  return Text(user.nickname, style: const TextStyle(fontSize: 20));
                 }
                 var data = snapshot.data!.data() as Map<String, dynamic>;
-                String nickname = data['nickname'] ?? user.nickname; // 서버에서 받아온 닉네임이 없으면 기본 닉네임 사용
+                String nickname = data['nickname'] ?? user.nickname;
                 return Text(nickname, style: const TextStyle(fontSize: 20));
               },
             ),
@@ -106,10 +105,9 @@ class _MyPageState extends State<MyPage> {
             ),
             ElevatedButton(
               onPressed: () {
-                // 수정 페이지로 이동
                 Navigator.push(
                   context,
-                  MaterialPageRoute(builder: (context) => UserData()), // 수정 페이지로 이동
+                  MaterialPageRoute(builder: (context) => UserData()),
                 );
               },
               child: const Text('내정보'),
@@ -124,34 +122,25 @@ class _MyPageState extends State<MyPage> {
     final XFile? pickedFile = await _picker.pickImage(source: ImageSource.gallery);
     if (pickedFile != null) {
       File imageFile = File(pickedFile.path);
-
-      // 파일 크기를 확인하지 않고 바로 이미지 업로드 수행
-      _uploadImage(imageFile);
+      await _uploadImage(imageFile);
+      setState(() {
+        isImageChanged = true;
+      });
     }
   }
 
-  void _uploadImage(File imageFile) async {
+  Future<void> _uploadImage(File imageFile) async {
     try {
-      // 현재 사용자의 이전 프로필 이미지 URL 가져오기
       String previousImageUrl = user.profileImageUrl;
-
-      // Firebase Storage에 새 이미지 업로드
       String uid = FirebaseAuth.instance.currentUser!.uid;
       Reference storageReference = FirebaseStorage.instance.ref().child('profile_images/$uid/${DateTime.now().millisecondsSinceEpoch}');
       UploadTask uploadTask = storageReference.putFile(imageFile);
       await uploadTask.whenComplete(() async {
-        // 업로드가 완료되면 새로운 이미지의 URL을 가져옴
         String imageUrl = await storageReference.getDownloadURL();
-
-        // Firestore에 새 이미지 URL 업데이트
         await FirebaseFirestore.instance.collection('UserData').doc(user.uid).update({'profileImageUrl': imageUrl});
-
-        // 이전 프로필 이미지가 있다면 삭제
-        if (previousImageUrl.isNotEmpty) {
+        if (previousImageUrl.isNotEmpty && previousImageUrl != defaultImageUrl) {
           await FirebaseStorage.instance.refFromURL(previousImageUrl).delete();
         }
-
-        // 프로필 이미지 경로를 업데이트
         setState(() {
           user.profileImageUrl = imageUrl;
         });
@@ -161,18 +150,15 @@ class _MyPageState extends State<MyPage> {
     }
   }
 
-  // 사용자 정보 불러오기
   void _loadUserData() async {
     try {
-      // Firestore에서 해당 사용자의 데이터 가져오기
       DocumentSnapshot userData = await FirebaseFirestore.instance.collection('UserData').doc(user.uid).get();
       if (userData.exists) {
         var data = userData.data() as Map<String, dynamic>;
-        String profileImageUrl = data['profileImageUrl'] ?? ''; // 프로필 이미지 URL이 없으면 빈 문자열 사용
-        String nickname = data['nickname'] ?? ''; // 서버에서 받아온 닉네임이 없으면 빈 문자열 사용
+        String profileImageUrl = data['profileImageUrl'] ?? defaultImageUrl;
+        String nickname = data['nickname'] ?? '';
         setState(() {
-          // 사용자 정보 업데이트
-          user.profileImageUrl = profileImageUrl;
+          user.profileImageUrl = profileImageUrl.isNotEmpty ? profileImageUrl : defaultImageUrl;
           user.nickname = nickname;
         });
       } else {
@@ -183,19 +169,16 @@ class _MyPageState extends State<MyPage> {
     }
   }
 
-  // 로그아웃 기능
   void _logout(BuildContext context) async {
     final GoogleSignIn googleSignIn = GoogleSignIn();
     try {
-      await googleSignIn.signOut(); // Google 계정에서 로그아웃
-      // FirebaseAuth에서도 로그아웃 수행
+      await googleSignIn.signOut();
       await FirebaseAuth.instance.signOut();
-      // 로그아웃 후 로그인 페이지로 이동
       Navigator.pushReplacement(
         context,
         MaterialPageRoute(builder: (context) => GoogleLogin()),
       );
-      print('로그아웃 성공'); // 로그아웃 성공 메시지 출력
+      print('로그아웃 성공');
     } catch (e) {
       print('로그아웃 중 오류 발생: $e');
     }
